+++ main.tf
@@ -85,6 +85,7 @@ data "template_file" "gitlab_runner" {
     gitlab_runner_version  = "${var.gitlab_runner_version}"
     docker_machine_version = "${var.docker_machine_version}"
     runners_config         = "${data.template_file.runners.rendered}"
+    runners_executor       = "${var.runners_executor}"
     pre_install            = "${var.userdata_pre_install}"
     post_install           = "${var.userdata_post_install}"
   }